Update LCD_Stopwatch
*This version "works", but the start/stop functionality using the 
SELECT button does not always work on the first try.  It kind of 
seems like you have to hold the button down for the perfect 
amount of time. 
* ALSO, the next version of this, which I am about to upload
as the next commit, is EXACTLY the same except for the Serial.print 
commands that a removed. HOWEVER, it seems to me that the Select
button's response is even worse after I removed the Serial.print 
commands. Perhaps they affect run time? I honestly have no idea
 what is going on with that!